+++ plugins/permission-node/src/ServerPermissionClient.test.ts
@@ -48,7 +48,9 @@ const discovery: PluginEndpointDiscovery = {
 };
 const testBasicPermission = createPermission({
   name: 'test.permission',
-  attributes: {},
+  attributes: {
+    action: 'create',
+  },
 });
 
 const testResourcePermission = createPermission({
@@ -361,5 +363,245 @@ describe('ServerPermissionClient', () => {
         );
       });
     });
+  });
+
+  describe('with access restrictions', () => {
+    it.each([{ enabled: true }, { enabled: false }])(
+      'short circuits the response when using a service principal, applying the relevant access restrictions if present, when permissions %p',
+      async permissionConfig => {
+        const client = ServerPermissionClient.fromConfig(
+          new ConfigReader({
+            permission: permissionConfig,
+          }),
+          {
+            discovery,
+            tokenManager: mockServices.tokenManager(),
+            auth: mockServices.auth(),
+          },
+        );
+
+        // no restrictions for the given plugin
+        await expect(
+          client.authorize(
+            [
+              {
+                permission: createPermission({
+                  name: 'test.permission',
+                  attributes: {
+                    action: 'create',
+                  },
+                }),
+              },
+            ],
+            {
+              credentials: mockCredentials.service('foo', {}),
+            },
+          ),
+        ).resolves.toEqual([{ result: AuthorizeResult.ALLOW }]);
+        await expect(
+          client.authorizeConditional(
+            [
+              {
+                resourceRef: undefined as any,
+                permission: createPermission({
+                  resourceType: 'test',
+                  name: 'test.permission',
+                  attributes: {
+                    action: 'create',
+                  },
+                }),
+              },
+            ],
+            {
+              credentials: mockCredentials.service('foo', {}),
+            },
+          ),
+        ).resolves.toEqual([{ result: AuthorizeResult.ALLOW }]);
+
+        // matching permission name
+        await expect(
+          client.authorize(
+            [
+              {
+                permission: createPermission({
+                  name: 'test.permission',
+                  attributes: {
+                    action: 'create',
+                  },
+                }),
+              },
+            ],
+            {
+              credentials: mockCredentials.service('foo', {
+                permissionNames: ['test.permission', 'other'],
+              }),
+            },
+          ),
+        ).resolves.toEqual([{ result: AuthorizeResult.ALLOW }]);
+        await expect(
+          client.authorizeConditional(
+            [
+              {
+                resourceRef: undefined as any,
+                permission: createPermission({
+                  resourceType: 'test',
+                  name: 'test.permission',
+                  attributes: {
+                    action: 'create',
+                  },
+                }),
+              },
+            ],
+            {
+              credentials: mockCredentials.service('foo', {
+                permissionNames: ['test.permission', 'other'],
+              }),
+            },
+          ),
+        ).resolves.toEqual([{ result: AuthorizeResult.ALLOW }]);
+
+        // matching attributes
+        await expect(
+          client.authorize(
+            [
+              {
+                permission: createPermission({
+                  name: 'test.permission',
+                  attributes: {
+                    action: 'create',
+                  },
+                }),
+              },
+            ],
+            {
+              credentials: mockCredentials.service('foo', {
+                permissionAttributes: {
+                  action: ['create', 'other' as any],
+                },
+              }),
+            },
+          ),
+        ).resolves.toEqual([{ result: AuthorizeResult.ALLOW }]);
+        await expect(
+          client.authorizeConditional(
+            [
+              {
+                resourceRef: undefined as any,
+                permission: createPermission({
+                  resourceType: 'test',
+                  name: 'test.permission',
+                  attributes: {
+                    action: 'create',
+                  },
+                }),
+              },
+            ],
+            {
+              credentials: mockCredentials.service('foo', {
+                permissionAttributes: {
+                  action: ['create', 'other' as any],
+                },
+              }),
+            },
+          ),
+        ).resolves.toEqual([{ result: AuthorizeResult.ALLOW }]);
+
+        // matching permission name but not attributes
+        await expect(
+          client.authorize(
+            [
+              {
+                permission: createPermission({
+                  name: 'test.permission',
+                  attributes: {
+                    action: 'create',
+                  },
+                }),
+              },
+            ],
+            {
+              credentials: mockCredentials.service('foo', {
+                permissionNames: ['test.permission'],
+                permissionAttributes: {
+                  action: ['other' as any],
+                },
+              }),
+            },
+          ),
+        ).resolves.toEqual([{ result: AuthorizeResult.DENY }]);
+        await expect(
+          client.authorizeConditional(
+            [
+              {
+                resourceRef: undefined as any,
+                permission: createPermission({
+                  resourceType: 'test',
+                  name: 'test.permission',
+                  attributes: {
+                    action: 'create',
+                  },
+                }),
+              },
+            ],
+            {
+              credentials: mockCredentials.service('foo', {
+                permissionNames: ['test.permission'],
+                permissionAttributes: {
+                  action: ['other' as any],
+                },
+              }),
+            },
+          ),
+        ).resolves.toEqual([{ result: AuthorizeResult.DENY }]);
+
+        // matching attributes but not permission name
+        await expect(
+          client.authorize(
+            [
+              {
+                permission: createPermission({
+                  name: 'test.permission',
+                  attributes: {
+                    action: 'create',
+                  },
+                }),
+              },
+            ],
+            {
+              credentials: mockCredentials.service('foo', {
+                permissionNames: ['wrong-name'],
+                permissionAttributes: {
+                  action: ['create'],
+                },
+              }),
+            },
+          ),
+        ).resolves.toEqual([{ result: AuthorizeResult.DENY }]);
+        await expect(
+          client.authorizeConditional(
+            [
+              {
+                resourceRef: undefined as any,
+                permission: createPermission({
+                  resourceType: 'test',
+                  name: 'test.permission',
+                  attributes: {
+                    action: 'create',
+                  },
+                }),
+              },
+            ],
+            {
+              credentials: mockCredentials.service('foo', {
+                permissionNames: ['wrong-name'],
+                permissionAttributes: {
+                  action: ['create'],
+                },
+              }),
+            },
+          ),
+        ).resolves.toEqual([{ result: AuthorizeResult.DENY }]);
+      },
+    );
   });
 });