+++ plugins/scaffolder/src/components/OngoingTask/OngoingTask.tsx
@@ -35,6 +35,12 @@ import {
   TaskSteps,
 } from '@backstage/plugin-scaffolder-react/alpha';
 import { useAsync } from '@react-hookz/web';
+import { usePermission } from '@backstage/plugin-permission-react';
+import {
+  taskCancelPermission,
+  taskReadPermission,
+  taskCreatePermission,
+} from '@backstage/plugin-scaffolder-common/alpha';
 
 const useStyles = makeStyles(theme => ({
   contentWrapper: {
@@ -81,6 +87,22 @@ export const OngoingTask = (props: {
 
   const [logsVisible, setLogVisibleState] = useState(false);
   const [buttonBarVisible, setButtonBarVisibleState] = useState(true);
+
+  // Used dummy string value for `resourceRef` since `allowed` field will always return `false` if `resourceRef` is `undefined`
+  const { allowed: canCancelTask } = usePermission({
+    permission: taskCancelPermission,
+  });
+
+  const { allowed: canReadTask } = usePermission({
+    permission: taskReadPermission,
+  });
+
+  const { allowed: canCreateTask } = usePermission({
+    permission: taskCreatePermission,
+  });
+
+  // Start Over endpoint requires user to have both read (to grab parameters) and create (to create new task) permissions
+  const canStartOver = canReadTask && canCreateTask;
 
   useEffect(() => {
     if (taskStream.error) {
@@ -197,7 +219,11 @@ export const OngoingTask = (props: {
                 <div className={classes.buttonBar}>
                   <Button
                     className={classes.cancelButton}
-                    disabled={!cancelEnabled || cancelStatus !== 'not-executed'}
+                    disabled={
+                      !cancelEnabled ||
+                      cancelStatus !== 'not-executed' ||
+                      !canCancelTask
+                    }
                     onClick={triggerCancel}
                     data-testid="cancel-button"
                   >
@@ -214,8 +240,9 @@ export const OngoingTask = (props: {
                   <Button
                     variant="contained"
                     color="primary"
-                    disabled={cancelEnabled}
+                    disabled={cancelEnabled || !canStartOver}
                     onClick={startOver}
+                    data-testid="start-over-button"
                   >
                     Start Over
                   </Button>