+++ plugins/example-todo-list-backend/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @internal/plugin-todo-list-backend
 
+## 1.0.28-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.19-next.1
+  - @backstage/backend-common@0.23.0-next.1
+  - @backstage/plugin-auth-node@0.4.14-next.1
+
 ## 1.0.28-next.0
 
 ### Patch Changes