+++ plugins/notifications-node/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @backstage/plugin-notifications-node
 
+## 0.2.0-next.1
+
+### Minor Changes
+
+- 07a789b: add notifications filtering by processors
+
+### Patch Changes
+
+- 1354d81: Use `node-fetch` instead of native fetch, as per https://backstage.io/docs/architecture-decisions/adrs-adr013
+- Updated dependencies
+  - @backstage/backend-plugin-api@0.6.19-next.1
+  - @backstage/backend-common@0.23.0-next.1
+  - @backstage/plugin-signals-node@0.1.5-next.1
+
 ## 0.1.5-next.0
 
 ### Patch Changes