+++ plugins/search-backend-module-techdocs/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @backstage/plugin-search-backend-module-techdocs
 
+## 0.1.24-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-tasks@0.5.24-next.1
+  - @backstage/backend-plugin-api@0.6.19-next.1
+  - @backstage/backend-common@0.23.0-next.1
+  - @backstage/plugin-catalog-node@1.12.1-next.0
+  - @backstage/plugin-search-backend-node@1.2.24-next.1
+  - @backstage/plugin-techdocs-node@1.12.5-next.1
+
 ## 0.1.24-next.0
 
 ### Patch Changes